Fixed vue router. Having issues with Vuetify.
