+++ package.json
@@ -16,7 +16,7 @@
     "primevue": "^3.34.0",
     "vue": "^3.2.13",
     "vue-router": "^4.2.4",
-    "vuetify": "^3.3.14"
+    "vuetify": "^3.3.15"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
