+++ src/App.vue
@@ -1,20 +1,8 @@
-<script setup>
-import { ref } from "vue";
-import Listbox from 'primevue/listbox';
-//import {useRouter} from "vue-router";
-import router from "../src/router.js";
-//const router = useRouter();
-//import ViewCourse from './components/addCourse.vue'
+<template>
+  <router-view />
+</template>
 
-const addCourse =() => {
-  console.log("help me plase");
-     router.push({ path: 'add'});
-       console.log("crying myself to sleep");
-
-
-};
-
-
+<script>
 // export default {
 //   name: 'App',
 //   components: {
@@ -35,26 +23,26 @@ const addCourse =() => {
 // });
 
 //One that works, ish
-const product = ref(null);
+//const product = ref(null);
 
 // POST request using fetch with set headers
-const requestOptions = {
-    method: 'POST',
-    headers: { 
-        'Content-Type': 'application/json',
-        'My-Custom-Header': 'foobar'
-    },
-    body: JSON.stringify({
-       courseNum: 'Test',
-       dept: 'test',
-       level: '1',
-       hours: '1',
-       name: 'test'
-       })
-};
-fetch('http://localhost:8081/', requestOptions)
-    .then(response => response.json())
-    .then(data => product.value = data);
+// const requestOptions = {
+//     method: 'POST',
+//     headers: { 
+//         'Content-Type': 'application/json',
+//         'My-Custom-Header': 'foobar'
+//     },
+//     body: JSON.stringify({
+//        courseNum: 'Test',
+//        dept: 'test',
+//        level: '1',
+//        hours: '1',
+//        name: 'test'
+//        })
+// };
+// fetch('http://localhost:8081/', requestOptions)
+//     .then(response => response.json())
+//     .then(data => product.value = data);
 
 // // POST request using fetch with set headers
 // import CourseServices from "../courseServices.js";
@@ -78,62 +66,5 @@ fetch('http://localhost:8081/', requestOptions)
 //     });
 
 
-// const selectedCity = ref();
-// const cities = ref([
-//     { name: 'Course', code: 'NY' },
-//     { name: 'Course', code: 'RM' },
-//     { name: 'Course', code: 'LDN' },
-//     { name: 'Course', code: 'IST' },
-//     { name: 'Course', code: 'PRS' },
-//     { name: 'Course', code: '3' },
-//     { name: 'Course', code: '4' }
-// ]);
 
 </script>
-
-
-
-
-<style>
-/* #app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
-} */
-</style>
-
-
-<template>
-    <!--MenuBar /-->
-    <!--ViewCourse /-->
-    <div style="background-color: maroon; width 100%; height:50px; margin:-10px; display:block;">
-      <h1 style="color:white; text-align:center; margin:0px; padding-top:10px;">View Courses</h1>
-      </div>
-
-      <div class="column">
-    <div class="card flex justify-content-center">
-        <Listbox v-model="selectedCity" :options="cities" filter optionLabel="name" />
-    </div>
-    <div style="margin-top: 7rem"> 
-      <h1 style="text-align: center;">Actions</h1>
-      <div class="row">
-      <button >View</button>
-      <button>Edit</button>
-      </div>
-      <div class="row">
-      <button @click=addCourse>Add</button>
-      <!-- <a href="../views/addCourses.vue">Add</a> -->
-
-      <button>Delete</button>
-      </div>
-      </div>
-    </div>
-    
-    <div class="card text-center m-3">
-        <div class="card-body">New course courseNum, maybe:</div>
-    </div>
-</template>
-
