+++ src/main.js
@@ -1,22 +1,20 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-//import vuetify from "./plugins/vuetify.js";
+import vuetify from "./plugins/vuetify.js";
 // import "primeflex/primeflex.css";
- import "primevue/resources/themes/lara-light-blue/theme.css";
- import "primevue/resources/primevue.min.css"; /* Deprecated */
+import "primevue/resources/themes/lara-light-blue/theme.css";
+import "primevue/resources/primevue.min.css"; /* Deprecated */
 import "primeicons/primeicons.css";
 import "./style.css";
 
-
-//import { router } from "./router";
 import PrimeVue from "primevue/config";
 import Listbox from 'primevue/listbox';
-
+import router from './router'
 
 
 const app = createApp(App);
 
-app.use(PrimeVue, { ripple: true  });
+app.use(router, vuetify, PrimeVue, { ripple: true  });
 
 app.component('list-box', Listbox);
