+++ src/views/addCourses.vue
@@ -1,4 +1,5 @@
-  <script>
+<script>
+  
   export default {
     data: () => ({
       valid: false,
@@ -87,3 +88,14 @@
       </v-container>
     </v-form>
   </template>
+
+<style>
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
+}
+</style>
